Moved LatinLexicon model into language service for latin.
