+++ lemmatization/models.py
@@ -38,24 +38,6 @@ class FormToLemma(models.Model):
             )
 
 
-class LatinLexicon(models.Model):
-    """
-    This model contains a copy of the data from LatinMorph16.db
-    and is intended to replace the external Perseids Morphology Service
-    for the purpose of headword/lemma retrieval.
-    """
-    token = models.CharField(max_length=255)
-    lemma = models.CharField(max_length=255)
-    rank = models.IntegerField(default=999999)
-    count = models.IntegerField(default=0)
-    rate = models.FloatField(default=0)
-
-    class Meta:
-        constraints = [
-            models.UniqueConstraint(fields=["token", "lemma"], name="unique_token_lemma")
-        ]
-
-
 def lookup_form(form, lang):
     entries = FormToLemma.objects.filter(form=form, lang=lang)
     return set(entries.values_list("lemma", flat=True))
